# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridCacheAdapter.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridCacheSwapEntryImpl.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/distributed/dht/GridDhtTxPrepareFuture.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/transactions/IgniteTxLocalAdapter.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/transactions/TxDeadlock.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/odbc/OdbcQueryExecuteRequest.java
#	modules/core/src/test/java/org/apache/ignite/internal/processors/cache/distributed/dht/GridCacheDhtEvictionsDisabledSelfTest.java

+++ modules/core/src/main/java/org/apache/ignite/plugin/extensions/communication/MessageWriter.java
@@ -48,9 +48,10 @@ public interface MessageWriter {
      * Writes message header.
      *
      * @param type Message type.
-     * @param fieldCnt Fields count.ß
+     * @param fieldCnt Fields count.
+     * @return {@code true} if successfully. Otherwise returns {@code false}.
      */
-    public boolean writeHeader(byte type, byte fieldCnt);
+    public boolean writeHeader(short type, byte fieldCnt);
 
     /**
      * Writes {@code byte} value.
